Also move recovery mode functions up to BasePool
